Much better tests
