+++ tests/bench.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-ali --duration=10s \
-    --rate=20 \
-    --workers=5 \
+ali --duration=5s \
+    --rate=35 \
+    --workers=12 \
     --header 'Content-Type: application/json' \
     --body-file=$1 \
     --method=POST http://localhost:9000/inference
